% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{conditional_frequency_fraction}
\alias{conditional_frequency_fraction}
\title{Create frequency plot}
\usage{
conditional_frequency_fraction(
  traj,
  coordinates,
  gritty,
  siteDF,
  fraction = 0.16,
  type = "above",
  remove.outliers = T
)
}
\arguments{
\item{traj}{dataframe of trajectory points}

\item{coordinates}{vector of latlon coordinates associated with trajectories}

\item{gritty}{blank grid of the extent of trajectories}

\item{siteDF}{}

\item{type}{}

\item{remove.outliers}{}
}
\description{
Calculates conditional frequency above or below a given
fraction
}
