% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traj.R
\name{get_traj}
\alias{get_traj}
\title{Returns data frame of HYSPLIT trajectory runs}
\usage{
get_traj(
  trajDate,
  dur = 8,
  OutDir,
  Lat = 38.104841,
  Lon = -122.245442,
  Dir = "backward",
  Met = "reanalysis",
  ExecDir = NULL,
  MetDir = NULL,
  ht = 1000
)
}
\arguments{
\item{trajDate}{character string of dates}

\item{dur}{the duration of each trajectory model run in hours}

\item{OutDir}{a file path to where the trajectory output files will be stored}

\item{Lat}{the starting latitude (in decimal degrees)}

\item{Lon}{the starting longitude (in decimal degrees)}

\item{Dir}{the direction of the trajectory model run. Options are: 'forward'
or 'backward'}

\item{Met}{an option to select meteorological data files. The options are
gdas1 (Global Data Assimilation System 1-degree resolution data),
reanalysis (NCAR/NCEP global reanalysis data), and narr (North American
Regional Reanalysis).}

\item{ExecDir}{an optional file path for the working directory of the model
input and output files.}

\item{MetDir}{an optional file path to store or find existing meteorological
data files}

\item{ht}{the starting height for the model run (in meters above ground
level)}
}
\value{
a dataframe of interpolated HYSPLIT trajectory points
}
\description{
The function returns a data frame of interpolated HYSPLIT
trajectories
}
