% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Creates a conditional frequency fraction map from a given region of Vallejo
and statistic}
\usage{
make_map(region, statistic, basemap, buffer = 10)
}
\arguments{
\item{region}{Region of Vallejo. Options are: North, South, Downtown,
Foothills, Mare Island}

\item{statistic}{Options are: Ninetieth, Fiftieth, Tenth, Mean}

\item{basemap}{ggmap basemap}

\item{buffer}{Fraction out of 100 to exclude.}
}
\description{
Creates a conditional frequency fraction map from a given region of Vallejo
and statistic
}
