% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_more_maps}
\alias{make_more_maps}
\title{Creates multiple conditional frequency fraction maps}
\usage{
make_more_maps(regions, statistics, basemap, buffer = 10)
}
\arguments{
\item{regions}{Regions of Vallejo. Options are: North, South, Downtown,
Foothills, Mare Island}

\item{statistics}{Options are: Ninetieth, Fiftieth, Tenth, Mean}

\item{basemap}{ggmap basemap}

\item{buffer}{Fraction out of 100 to exclude.}
}
\description{
Creates multiple conditional frequency fraction maps
}
