% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{conditional_frequency_fraction}
\alias{conditional_frequency_fraction}
\title{Create frequency plot}
\usage{
conditional_frequency_fraction(
  traj,
  coordinates,
  gritty,
  siteDF,
  fraction = 0.16,
  type = "above",
  remove.outliers = T
)
}
\arguments{
\item{traj}{Data frame of trajectory points}

\item{coordinates}{Vector of latlon coordinates associated with trajectories}

\item{gritty}{Blank grid of the extent of trajectories}

\item{siteDF}{Data frame containing the emission concentrations and
coordinates}

\item{fraction}{Number between 0 and 1 to calculate conditional frequency
above or below}

\item{type}{Whether to calculate above or below the fraction}

\item{remove.outliers}{Removes all points above or below the fraction.
Default set to true.}
}
\description{
Calculates conditional frequency above or below a given fraction
}
